+++ plugins/amazon.py
@@ -1,6 +1,11 @@
 import mambase
 import utils
 
+
+def parse_and_format_date_german(datestr):
+    return utils.parse_and_format_date(datestr, languages=["de", "en"])
+
+
 class AmazonThirdParty(mambase.RegexMetaDataCheck):
     __documentclass__ = "G - Rechnung"
     __tags__ = ["Amazon"]
@@ -10,14 +15,14 @@ class AmazonThirdParty(mambase.RegexMetaDataCheck):
         {
             "metadata": "invoicedate",
             "regex": "Rechnungsdatum\n/Lieferdatum\n+(.*)",
-            "post": utils.parse_and_format_date,
+            "post": parse_and_format_date_german,
         },
         {"metadata": "invoicenumber", "regex": "Rechnungsnummer\n+(.*)",},
         {"metadata": "amount", "regex": "Zahlbetrag\n+(.*) €",},
         {
             "metadata": "orderdate",
             "regex": "Bestelldatum\n+(.*)",
-            "post": utils.parse_and_format_date,
+            "post": parse_and_format_date_german,
         },
         {"metadata": "ordernumber", "regex": "Bestellnummer\n+(.*)",},
     ]
